+++ src/taurenmd/cli_distances.py
@@ -46,8 +46,9 @@ import numpy as np
 from taurenmd import _BANNER, Path
 from taurenmd import core as tcore
 from taurenmd import log
-from taurenmd.libs import libcli, libio, libmda, libplot  # noqa: F401
+from taurenmd.libs import libcli, libio, libmda
 from taurenmd.logger import S, T
+from taurenmd.plots import plotparams
 
 
 __doc__ += (
@@ -72,7 +73,7 @@ libcli.add_insort_arg(ap)
 ap.add_argument(
     '-l1',
     '--sel1',
-    help='First selection.',
+    help='First selection (a single selection).',
     default='all',
     type=str,
     )
@@ -80,7 +81,11 @@ ap.add_argument(
 ap.add_argument(
     '-l2',
     '--sel2',
-    help='Second selection.',
+    help=(
+        'Second selections. As many as desired. For example: --sel2 "segid A" '
+        '"segid B and name CA".'
+        ),
+    nargs='*',
     default='all',
     type=str,
     )
@@ -124,62 +129,94 @@ def main(
 
     log.info(T('defining atom seletions'))
     log.info(S('atom selection #1: {}', sel1))
-    log.info(S('atom selection #2: {}', sel2))
+    log.info(S('atom selection #2: {}', ', '.join(sel2)))
     atom_sel1 = u.select_atoms(sel1)
-    atom_sel2 = u.select_atoms(sel2)
+    atom_sels2 = [u.select_atoms(_sel) for _sel in sel2]
 
-    distances = np.ones(len(u.trajectory[frame_slice]), dtype=np.float32)
+    distances = np.empty(
+        (len(u.trajectory[frame_slice]), len(sel2)),
+        dtype=np.float64,
+        )
 
     log.info(T('Calculating distances'))
     # https://www.mdanalysis.org/MDAnalysisTutorial/atomgroups.html
     # https://www.mdanalysis.org/docs/documentation_pages/core/groups.html#MDAnalysis.core.groups.AtomGroup.center_of_geometry
 
-    with libcli.ProgressBar(distances.size, suffix='frames') as pb:
+    with libcli.ProgressBar(distances.shape[0], suffix='frames') as pb:
         for i, _ts in enumerate(u.trajectory[frame_slice]):
+            for j, _sel2 in enumerate(sel2):
 
-            distances[i] = np.linalg.norm(
-                np.subtract(
-                    atom_sel1.center_of_geometry(),
-                    atom_sel2.center_of_geometry(),
+                distances[i, j] = np.linalg.norm(
+                    np.subtract(
+                        atom_sel1.center_of_geometry(),
+                        atom_sels2[j].center_of_geometry(),
+                        )
                     )
-                )
+
             pb.increment()
 
-    log.info(S('calculated a total of {} distances.', len(distances)))
+    log.info(S('calculated a total of {} distances.', distances.shape[0]))
 
     if export:
         libio.export_data_to_file(
             list(range(len(u.trajectory))[frame_slice]),
-            distances,
+            *[distances[:, d] for d in range(distances.shape[1])],
             fname=export,
             header=(
                 '# Distances between two selections centers of geomemtry\n'
                 '# topology: {}\n'
                 '# trajectories: {}\n'
                 '# selection #1: {}\n'
-                '# selection #2: {}\n'
-                '# frame,distance\n'
+                '# frame,{}\n'
                 ).format(
                     topology,
                     ', '.join(t.resolve().str() for t in trajectories),
                     sel1,
-                    sel2,
+                    ', '.join(sel2),
                     ),
             )
 
+    #if plot:
+    #    plotvars = plotvars or dict()
+    #    plotvars.setdefault('labels', '{} dist {}'.format(sel1, sel2))
+
+    #    log.info(T('plot params:'))
+    #    for k, v in plotvars.items():
+    #        log.info(S('{} = {!r}', k, v))
+
+    #    libplot.param(
+    #        list(range(len(u.trajectory))[frame_slice]),
+    #        distances,
+    #        **plotvars,
+    #        )
     if plot:
-        plotvars = plotvars or dict()
-        plotvars.setdefault('labels', '{} dist {}'.format(sel1, sel2))
+        log.info(T("Plotting results:"))
 
-        log.info(T('plot params:'))
-        for k, v in plotvars.items():
-            log.info(S('{} = {!r}', k, v))
+        subtitle = 'Main selection: {}'.format(sel1)
+        ymax = np.max(distances)
+        ymin = np.min(distances)
+        filename = 'plot_distances.pdf'
 
-        libplot.param(
+        cli_defaults = {
+            'ymax': ymax * 1.1 if ymax > 0 else ymax * 0.9,
+            'ymin': ymin * 1.1 if ymin < 0 else ymin * 0.9,
+            'filename': filename,
+            'title': f'Distance between a main selectio and others\n{subtitle}',
+            'xlabel': 'Frames',
+            'ylabel': r'Distance ($\AA$)',
+            'labels': sel2,
+            }
+
+        cli_defaults.update(plotvars or dict())
+
+        plotparams.plot(
             list(range(len(u.trajectory))[frame_slice]),
-            distances,
-            **plotvars,
+            [distances[:, d] for d in range(distances.shape[1])],
+            legend=True,
+            **cli_defaults,
             )
+
+        log.info(S(f'saved plot: {cli_defaults["filename"]}'))
 
     log.info(S('done'))
     return