+++ CHANGELOG.rst
@@ -10,6 +10,9 @@ Upon version 0.8, and before version 1, SV2 major version increments are reflect
 Changelog
 =========
 
+v0.10.2 (2021-12-03)
+------------------------------------------------------------
+
 * Add frame timestep and whole trajectory duration to `report` (#61)
 
 v0.10.1 (2021-11-24)